Move and improve price tests.
